nontriv nulli test: put non-colobbered ops >72h
the sequence of ops used by `log_nullification_nontrivial.json`,

```
(op0)<-(op1)<-(op2)<-(op3)<-(op4)  (op5)<-(op6)  (op7)
            \             \       /             /
             \             '--<--'             /
              '-------<------------------<----'
```

has *all* ops happening within 72h, but it seems clear that

- op0 can be arbitrarily earlier, and
- op1 can maybe also be arbirarily earlier? (at least that's what i'd expect)

this change moves op0 and op1 outside the 72h window, so that only *clobbered* events are inside it

the [wording in the spec](https://github.com/did-method-plc/did-method-plc?tab=readme-ov-file#key-rotation--account-recovery) about the conditions for clobbering ops was slightly unclear to me, so i tried to dig into test cases for certainty before suggesting a wording change, and ended up here. accepting or resolving this change (specifically re op1) would clarify my question about the spec.

---

note: i *expected* the test to fail when i moved `op2` outside the 72h window as well, but it did not. from what i can see, the tests only asserts that a sequence of events is made up of a valid sequence of submitted ops, but not that any are accepted/rejected or that any particular state results at the end. if that's not meant to be the case, i'd love to add assertions about that!